Rename AES syscall files to be clearer that they are H1B specific.
